자신의 상품아닐경우 삭제 불가능 기능 추가, 판매자 이름정보 상품정보에 추가

+++ src/main/java/com/shopingmall/seungjae/controller/Item/ShoppingMallItemController.java
@@ -1,6 +1,8 @@
 package com.shopingmall.seungjae.controller.Item;
 
+import com.shopingmall.seungjae.controller.Member.SessionConst;
 import com.shopingmall.seungjae.domain.Item;
+import com.shopingmall.seungjae.domain.Member;
 import com.shopingmall.seungjae.repository.ItemRepository;
 import com.shopingmall.seungjae.repository.ItemRepositoryImpl;
 import jakarta.validation.Valid;
@@ -37,8 +39,7 @@ public class ShoppingMallItemController {
     }
 
     @PostMapping("/add")
-    public String AddItem(@ModelAttribute Item item, BindingResult bindingResult) {
-        item.setSellerName("me");
+    public String AddItem(@ModelAttribute Item item, BindingResult bindingResult, @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember) {
         if (!StringUtils.hasText(item.getItemName())){
             bindingResult.addError(new FieldError("items","itemName", "상품 이름이 올바르지 않습니다."));
         }
@@ -52,6 +53,7 @@ public class ShoppingMallItemController {
             log.info("상품 추가 오류 발생 bindingResult = {}",bindingResult);
             return "item/addItemForm";
         }
+        item.setSellerName(loginMember.getName());
         itemRepository.save(item);
         log.info("item add ={}",item);
         return "redirect:/items";
@@ -67,12 +69,20 @@ public class ShoppingMallItemController {
     }
 
     @PostMapping("/item/{itemName}")
-    public String ItemDelete(@PathVariable(name = "itemName") String findItemName, Model model) {
+    public String ItemDelete(@PathVariable(name = "itemName") String findItemName, Model model,
+                             @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember) {
         Optional<Item> item = itemRepository.findByName(findItemName);
         Item itemNotNull = item.orElse(null); // Optional 객체 언랩
-        log.info("deleteItem = {}",itemNotNull);
-        itemRepository.delete(itemNotNull.getItemId());
-        return "redirect:/items";
+        if(loginMember.getName().equals(itemNotNull.getSellerName())&&loginMember!=null&&itemNotNull!=null) {
+            log.info("deleteItem = {}", itemNotNull);
+            itemRepository.delete(itemNotNull.getItemId());
+            return "redirect:/items";
+        }
+        else{
+            model.addAttribute("item", itemNotNull);
+            model.addAttribute("deleteError","본인의 제품이 아닐경우 삭제하실 수 없습니다.");
+            return "item/item";
+        }
     }
 
 }
